Remove white background from `.pagination-inner`
The header / footer of the pages have a white background set in the `.pagination-inner` class. The "body" of the pages however have no background color set and are transparent.

This results in the document looking "broken" if SuperDoc is used on anything but a white background. Hence this commit removing the explicit white background and leave the whole page background as transparent.