feat: store resolved numbering properties
The top-level attributes in the paragraph node are used to store the
resolved values, whereas the ones inside the "paragraphProperties"
attribute are the inline ones. The numbering properties are now being
stored as a top-level attribute too so that if the list's numId and
level are defined in a style, and not inline, they are correctly
accounted for.