refactor: parse numbering properties from converter's object
The logic used for resolving a paragraph's properties used to parse the
numbering definitions from the docx. That was changed so that the same
object that's used by the list helper functions is used. That way, the
existing logic related to adding a missing numbering definition will
work without changes.